Substitute InDels into one or more sequences

Modify one or more sequences to include Insertions or Deletions

armIndello(x, indels, ...)

# S4 method for XString,GRanges
armIndello(x, indels, exons, alt_col = "ALT", ...)

# S4 method for DNAStringSet,GRanges
armIndello(x, indels, alt_col = "ALT", mc_cores = 1, ...)

# S4 method for BSgenome,GRanges
armIndello(x, indels, alt_col = "ALT", mc_cores = 1, names, ...)

Arguments

x

Sequence of class XString

indels

GRanges object with InDel locations and the alternate allele

...

Not used

exons

GRanges object containing exon structure for x

alt_col

Column containing the alternate allele

mc_cores

Number of cores to use when calling parallel::mclapply internally

names

passed to BSgenome::getSeq when x is a BSgenome object

Value

A DNAStringSet or XString object (See Details)

Details

Takes an Biostrings::XString or Biostrings::XStringSet object and moddfies the sequence to incorporate InDels. The expected types of data determine the behaviour, with the following expectations describing how the function will incorporate data

Input Data Type	Exons Required	Use Case	Returned

XString	Y	Modify a Reference Transcriptome	XString

DNAStringSet	N	Modify a Reference Genome	DNAStringSet

BSgenome	N	Modify a Reference Genome	DNAStringSet

See also

Examples

## Start with a DNAStringSet
library(GenomicRanges)
seq <- DNAStringSet(c(seq1 = "AATCTGCGC"))
## Define an Insertion
var <- GRanges("seq1:1")
var$ALT <- "AAA"
seq
#> DNAStringSet object of length 1:
#>     width seq                                               names               
#> [1]     9 AATCTGCGC                                         seq1
armIndello(seq, var)
#> Updating seq1; Original length: 9
#> ; Updated length: 11
#> DNAStringSet object of length 1:
#>     width seq                                               names               
#> [1]    11 AAAATCTGCGC                                       seq1

## To modify a single transcript
library(GenomicFeatures)
#> Loading required package: AnnotationDbi
ex <- GRanges(c("seq1:1-3:+", "seq1:7-9:+"))
orig <- extractTranscriptSeqs(seq, GRangesList(tx1 = ex))[["tx1"]]
orig
#> 6-letter DNAString object
#> seq: AATCGC
armIndello(orig, var, exons = ex)
#> 8-letter DNAString object
#> seq: AAAATCGC