Count overlaps by variant type

Count how many variants overlap each element of a GRangesList

overlapsByVar(x, var, alt_col = "ALT", ...)

Arguments

x

A GRangesList with features of interest

var

A Granges object with variants of interest

alt_col

The column within mcols(var) which contains the alternate allele

...

Passed to rowSums

Value

A data.frame

Details

Taking any GRangesList, count how many of each variant type overlap a region within each element.

Examples

library(rtracklayer)
library(VariantAnnotation)
gtf <- import.gff(
    system.file("extdata/gencode.v44.subset.gtf.gz", package = "transmogR")
)
grl <- splitAsList(gtf, gtf$type)
vcf <- system.file("extdata/1000GP_subset.vcf.gz", package = "transmogR")
var <- rowRanges(readVcf(vcf, param = ScanVcfParam(fixed = "ALT")))
overlapsByVar(grl, var)
#>            Deletion Insertion SNV Total
#> gene              6         1  91    98
#> transcript        6         1  91    98
#> exon              0         1  37    38