Substitute InDels into one or more sequences

Modify one or more sequences to include Insertions or Deletions

armIndello(x, indels, exons, ...)

# S4 method for XString,GRanges,GRanges
armIndello(x, indels, exons, alt_col = "ALT", ...)

# S4 method for XStringSet,GRanges,GRanges
armIndello(x, indels, exons, alt_col = "ALT", ...)

# S4 method for DNAStringSet,GRanges,missing
armIndello(x, indels, exons, alt_col = "ALT", ..., BPPARAM = SerialParam())

# S4 method for BSgenome,GRanges,missing
armIndello(x, indels, exons, alt_col = "ALT", ..., BPPARAM = SerialParam())

Arguments

x

Sequence of class XString

indels

GRanges object with InDel locations and the alternate allele

exons

GRanges object containing exon structure for x

...

Not used

alt_col

Column containing the alternate allele

BPPARAM

BiocParallel::BiocParallelParam instance

Details

Takes an Biostrings::XString or Biostrings::XStringSet object and moddfies the sequence to incorporate InDels. The expected types of data determine the behaviour, with the following expectations describing how the function will incorporate data

Input Data Type	Exons Required	Use Case	Returned

XString	Y	Modify a Reference Transcriptome	XString

XStringSet	Y	Modify a Reference Transcriptome	XStringSet

DNAStringSet	N	Modify a Reference Genome	DNAStringSet

BSgenome	N	Modify a Reference Genome	DNAStringSet