Obtain Splice-Juntions from Exons and Transcripts

Using GRanges defining exons and transcripts, find the splice-junctions

sjFromExons(
  x,
  rank_col = c("exon_number", "exon_rank"),
  tx_col = c("transcript_id", "tx_id"),
  extra_cols = "all",
  don_len = 8,
  acc_len = 5,
  chop = TRUE,
  ...
)

Arguments

x

GRanges object with exons and transcripts. A column indicating the position (or rank) of each exon within the transcript must be included.

rank_col

The column containing the position of each exons within the transcript

tx_col

The column containing unique transcript-level identifiers

extra_cols

Can be a vector of column names to return beyond rank_col and tx_col. By default all columns are returned (extra_cols = "all").

don_len, acc_len

Length of donor and acceptor sites respectively

chop

logical(1) Apply extraChIPs::chopMC to the output to only return unique ranges, with all mcols collapsed.

...

Passed to extraChIPs::chopMC

Value

A GRanges object with requested columns, and an additional column, 'site', annotating each region as a donor or acceptor site.

Details

A canonical splice junction consists of a donor site and an acceptor site at each end of an intron, with a branching site somewhere wthin the intron. Canonical donor sites are 8nt long with the the first two bases being exonic and the next 6 being derived form intronic sequences. Canonical acceptor sites are 5nt long with the first four bases being intronic and the final base being the first base of the next exon.

This functions uses each set of exons within a transcript to identify both donor and acceptor sites. Branch sites are not identified.

Examples

library(rtracklayer)
gtf <- import.gff(
   system.file("extdata/gencode.v44.subset.gtf.gz", package = "transmogR"),
   feature.type = "exon"
)
sjFromExons(gtf)
#> GRanges object with 196 ranges and 20 metadata columns:
#>         seqnames        ranges strand |        site
#>            <Rle>     <IRanges>  <Rle> | <character>
#>     [1]     chr1 182745-182752      + |       donor
#>     [2]     chr1 183128-183132      + |    acceptor
#>     [3]     chr1 183215-183222      + |       donor
#>     [4]     chr1 183490-183494      + |    acceptor
#>     [5]     chr1 183570-183577      + |       donor
#>     ...      ...           ...    ... .         ...
#>   [192]     chr1 808623-808627      - |    acceptor
#>   [193]     chr1 810061-810068      - |       donor
#>   [194]     chr1 810170-810174      - |    acceptor
#>   [195]     chr1 817367-817374      - |       donor
#>   [196]     chr1 827664-827671      - |       donor
#>                               transcript_id        source     type     score
#>                             <CharacterList>  <FactorList> <factor> <numeric>
#>     [1]                   ENST00000624431.2        HAVANA     exon        NA
#>     [2]                   ENST00000624431.2        HAVANA     exon        NA
#>     [3]                   ENST00000624431.2        HAVANA     exon        NA
#>     [4]                   ENST00000624431.2        HAVANA     exon        NA
#>     [5]                   ENST00000624431.2        HAVANA     exon        NA
#>     ...                                 ...           ...      ...       ...
#>   [192] ENST00000447500.4,ENST00000590817.3 HAVANA,HAVANA     exon        NA
#>   [193] ENST00000447500.4,ENST00000590817.3 HAVANA,HAVANA     exon        NA
#>   [194]                   ENST00000447500.4        HAVANA     exon        NA
#>   [195]                   ENST00000447500.4        HAVANA     exon        NA
#>   [196] ENST00000635509.2,ENST00000634337.2 HAVANA,HAVANA     exon        NA
#>             phase                               gene_id
#>         <integer>                       <CharacterList>
#>     [1]      <NA>                     ENSG00000279928.2
#>     [2]      <NA>                     ENSG00000279928.2
#>     [3]      <NA>                     ENSG00000279928.2
#>     [4]      <NA>                     ENSG00000279928.2
#>     [5]      <NA>                     ENSG00000279928.2
#>     ...       ...                                   ...
#>   [192]      <NA>   ENSG00000290784.1,ENSG00000230092.8
#>   [193]      <NA>   ENSG00000290784.1,ENSG00000230092.8
#>   [194]      <NA>                     ENSG00000290784.1
#>   [195]      <NA>                     ENSG00000290784.1
#>   [196]      <NA> ENSG00000230021.10,ENSG00000230021.10
#>                                             gene_type
#>                                       <CharacterList>
#>     [1]                        unprocessed_pseudogene
#>     [2]                        unprocessed_pseudogene
#>     [3]                        unprocessed_pseudogene
#>     [4]                        unprocessed_pseudogene
#>     [5]                        unprocessed_pseudogene
#>     ...                                           ...
#>   [192]                 lncRNA,transcribed_unproces..
#>   [193]                 lncRNA,transcribed_unproces..
#>   [194]                                        lncRNA
#>   [195]                                        lncRNA
#>   [196] transcribed_processe..,transcribed_processe..
#>                               gene_name           level
#>                         <CharacterList> <CharacterList>
#>     [1]                        DDX11L17               2
#>     [2]                        DDX11L17               2
#>     [3]                        DDX11L17               2
#>     [4]                        DDX11L17               2
#>     [5]                        DDX11L17               2
#>     ...                             ...             ...
#>   [192] ENSG00000290784,ENSG00000230092             2,2
#>   [193] ENSG00000290784,ENSG00000230092             2,2
#>   [194]                 ENSG00000290784               2
#>   [195]                 ENSG00000290784               2
#>   [196] ENSG00000230021,ENSG00000230021             2,2
#>                                         tag
#>                             <CharacterList>
#>     [1]                   Ensembl_canonical
#>     [2]                   Ensembl_canonical
#>     [3]                   Ensembl_canonical
#>     [4]                   Ensembl_canonical
#>     [5]                   Ensembl_canonical
#>     ...                                 ...
#>   [192] Ensembl_canonical,Ensembl_canonical
#>   [193] Ensembl_canonical,Ensembl_canonical
#>   [194]                   Ensembl_canonical
#>   [195]                   Ensembl_canonical
#>   [196]        RNA_Seq_supported_only,basic
#>                                   transcript_type
#>                                   <CharacterList>
#>     [1]                    unprocessed_pseudogene
#>     [2]                    unprocessed_pseudogene
#>     [3]                    unprocessed_pseudogene
#>     [4]                    unprocessed_pseudogene
#>     [5]                    unprocessed_pseudogene
#>     ...                                       ...
#>   [192]             lncRNA,transcribed_unproces..
#>   [193]             lncRNA,transcribed_unproces..
#>   [194]                                    lncRNA
#>   [195]                                    lncRNA
#>   [196] processed_transcript,processed_transcript
#>                         transcript_name transcript_support_level
#>                         <CharacterList>          <CharacterList>
#>     [1]                    DDX11L17-201                       NA
#>     [2]                    DDX11L17-201                       NA
#>     [3]                    DDX11L17-201                       NA
#>     [4]                    DDX11L17-201                       NA
#>     [5]                    DDX11L17-201                       NA
#>     ...                             ...                      ...
#>   [192] ENST00000447500,ENST00000590817                     5,NA
#>   [193] ENST00000447500,ENST00000590817                     5,NA
#>   [194]                 ENST00000447500                        5
#>   [195]                 ENST00000447500                        5
#>   [196] ENST00000635509,ENST00000634337                      5,5
#>                                 havana_transcript
#>                                   <CharacterList>
#>     [1]                      OTTHUMT00000491420.1
#>     [2]                      OTTHUMT00000491420.1
#>     [3]                      OTTHUMT00000491420.1
#>     [4]                      OTTHUMT00000491420.1
#>     [5]                      OTTHUMT00000491420.1
#>     ...                                       ...
#>   [192] OTTHUMT00000448550.2,OTTHUMT00000006886.2
#>   [193] OTTHUMT00000448550.2,OTTHUMT00000006886.2
#>   [194]                      OTTHUMT00000448550.2
#>   [195]                      OTTHUMT00000448550.2
#>   [196] OTTHUMT00000488653.2,OTTHUMT00000488647.3
#>                                     exon_id
#>                             <CharacterList>
#>     [1]                   ENSE00003759020.2
#>     [2]                   ENSE00003759581.2
#>     [3]                   ENSE00003759581.2
#>     [4]                   ENSE00003804405.1
#>     [5]                   ENSE00003804405.1
#>     ...                                 ...
#>   [192] ENSE00004015917.1,ENSE00001782999.1
#>   [193] ENSE00001674926.2,ENSE00001651951.1
#>   [194]                   ENSE00001674926.2
#>   [195]                   ENSE00001746491.1
#>   [196] ENSE00003786980.2,ENSE00003786823.2
#>                                       havana_gene     hgnc_id             ont
#>                                   <CharacterList> <character> <CharacterList>
#>     [1]                      OTTHUMG00000191962.1  HGNC:55080     PGO:0000005
#>     [2]                      OTTHUMG00000191962.1  HGNC:55080     PGO:0000005
#>     [3]                      OTTHUMG00000191962.1  HGNC:55080     PGO:0000005
#>     [4]                      OTTHUMG00000191962.1  HGNC:55080     PGO:0000005
#>     [5]                      OTTHUMG00000191962.1  HGNC:55080     PGO:0000005
#>     ...                                       ...         ...             ...
#>   [192]                   NA,OTTHUMG00000002403.3        <NA>  NA,PGO:0000019
#>   [193]                   NA,OTTHUMG00000002403.3        <NA>  NA,PGO:0000019
#>   [194]                                      <NA>        <NA>            <NA>
#>   [195]                                      <NA>        <NA>            <NA>
#>   [196] OTTHUMG00000191652.4,OTTHUMG00000191652.4        <NA>           NA,NA
#>             exon_number
#>         <CharacterList>
#>     [1]               1
#>     [2]               2
#>     [3]               2
#>     [4]               3
#>     [5]               3
#>     ...             ...
#>   [192]             3,2
#>   [193]             2,1
#>   [194]               2
#>   [195]               1
#>   [196]             1,1
#>   -------
#>   seqinfo: 1 sequence from an unspecified genome; no seqlengths